requirements.txt
streamlit>=1.36
python-dotenv>=1.0.0
requests>=2.31
